Fix #192: tiff files (and others) appear blurry when 1:1 view
https://github.com/BestImageViewer/geeqie/issues/192

Up to commit 808d8558b3d5 alpha-channel files at scale 1:1 always had
GdkInterpType set to GDK_INTERP_NEAREST - the lowest quality. As there
was no scaling, this was ok.
Thereafter scaling was set to that selected by the user, resulting in
problems when GDK_INTERP_HYPER was selected.
This commit reverts to the previous method.